+++ src/main/java/com/proyecto/service/TrabajadorService.java
@@ -10,8 +10,8 @@ public interface TrabajadorService {
 	//public Trabajador iniciarSesion(String usuario, String clave);
 	public Trabajador registrar(Trabajador trabajador);
 	public Trabajador actualizar(Trabajador trabajador);
-	public Integer eliminar(int idtra);
+	public Integer eliminar(Long idtra);
 	public List<Trabajador>listar();	
-	Optional<Trabajador>obtenerPorId(int idtra);
+	Optional<Trabajador>obtenerPorId(Long idtra);
 	public Trabajador findByUsername(String username);
 }
